# perf(http): avoid Set.has() when closing connection resource