--- a/ext/http/01_http.js
+++ b/ext/http/01_http.js
@@ -182,6 +182,7 @@
     localAddr,
   ) {
     return async function respondWith(resp) {
+      let closeQuick = false;
       try {
         if (ObjectPrototypeIsPrototypeOf(PromisePrototype, resp)) {
           resp = await resp;
@@ -241,15 +242,12 @@
           ObjectPrototypeIsPrototypeOf(Uint8ArrayPrototype, respBody)
         );
 
-        const deferred = request[_deferred];
-        const ws = resp[_ws];
-        
+
         try {
           await core.opAsync(
             "op_http_write_headers",
             [streamRid, innerResp.status ?? 200, innerResp.headerList],
             isStreamingResponseBody ? null : respBody,
-            ws && deferred && isStreamingResponseBody,
           );
         } catch (error) {
           const connError = httpConn[connErrorSymbol];
@@ -307,6 +305,7 @@
           }
         }
 
+        const deferred = request[_deferred];
         if (deferred) {
           const res = await core.opAsync("op_http_upgrade", streamRid);
           let conn;
@@ -322,6 +321,8 @@
 
           deferred.resolve([conn, res.readBuf]);
         }
+
+        const ws = request[_ws];
         if (ws) {
           const wsRid = await core.opAsync(
             "op_http_upgrade_websocket",
@@ -360,10 +361,10 @@
           }
         }
       } finally {
-        if (SetPrototypeHas(httpConn.managedResources, streamRid)) {
-          SetPrototypeDelete(httpConn.managedResources, streamRid);
-          core.close(streamRid);
-        }
+         const deleted = SetPrototypeDelete(httpConn.managedResources, streamRid);
+         if(deleted) {
+           core.close(streamRid)
+         }
       }
     };
   }