--- a/ext/http/01_http.js
+++ b/ext/http/01_http.js
@@ -181,7 +181,6 @@
     localAddr,
   ) {
     return async function respondWith(resp) {
-      let closeQuick = false;
       try {
         resp = await resp;
         if (!(ObjectPrototypeIsPrototypeOf(ResponsePrototype, resp))) {