--- a/ext/http/01_http.js
+++ b/ext/http/01_http.js
@@ -42,7 +42,6 @@
     Set,
     SetPrototypeAdd,
     SetPrototypeDelete,
-    SetPrototypeHas,
     SetPrototypeValues,
     StringPrototypeIncludes,
     StringPrototypeToLowerCase,